requirements.txt
langchain-openai
langchain_community
langchain-huggingface
python-dotenv
langchain-text-splitters
langchain-chroma